refactor(test/perf): host native scenarios outside perfcheck
